Separate getMetricsAppModelMethods
